代码优化

+++ Tools/ConfigDataExpoter/ConfigDataExpoter/FileExporter.cs
@@ -9,16 +9,16 @@ namespace ConfigDataExpoter
 {
     class FileExporter
     {
-        public void ExportFile(string filePath, string code)
+        public void ExportFile(string filePath, string code, bool recreateDirectory = true)
         {
             var directoryName = Path.GetDirectoryName(filePath);
             try
             {
-                if (Directory.Exists(directoryName))
+                if (Directory.Exists(directoryName) && recreateDirectory)
                 {
                     Directory.Delete(directoryName, true);
+                    Directory.CreateDirectory(directoryName);
                 }
-                Directory.CreateDirectory(directoryName);
                 var streamWriter = File.CreateText(filePath);
                 streamWriter.Write(code);
                 streamWriter.Flush();
